hotfixes for influx; making the write API shared
